requirements.txt
typer==0.12.*
tqdm==4.*
pydantic==2.*
pyyaml==6.*
opencv-python==4.*
Pillow==10.*
numpy==1.*
onnxruntime-gpu==1.19.*
sentencepiece==0.2.*
lxml==5.*
ujson==5.*
